xbps-src: fix previous correctly for local mode too.

+++ common/xbps-src/shutils/chroot.sh
@@ -136,14 +136,18 @@ chroot_sync_repos() {
             fi
             cp -a $XBPS_MASTERDIR/var/db/xbps/keys/*.plist \
                 $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/var/db/xbps/keys
-            rm -rf $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
-            mkdir -p $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
-            cp ${XBPS_MASTERDIR}/etc/xbps/repo.d/*.conf \
-                $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
             env XBPS_TARGET_ARCH=$XBPS_TARGET_ARCH \
                 $CHROOT_CMD $XBPS_MASTERDIR /usr/sbin/xbps-install \
                 -r /usr/$XBPS_CROSS_TRIPLET -S
         fi
+    fi
+
+    # Copy host repos to the cross root.
+    if [ -n "$XBPS_CROSS_BUILD" ]; then
+        rm -rf $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
+        mkdir -p $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
+        cp ${XBPS_MASTERDIR}/etc/xbps/repo.d/*.conf \
+            $XBPS_MASTERDIR/usr/$XBPS_CROSS_TRIPLET/etc/xbps/repo.d
     fi
 
     return 0
